+++ Content.Shared/Fluids/SharedPuddleSystem.cs
@@ -1,5 +1,4 @@
 using System.Linq;
-using Content.Shared._Funkystation.Fluids;
 using Content.Shared.Administration.Logs;
 using Content.Shared.Chemistry;
 using Content.Shared.Chemistry.Components;
@@ -11,22 +10,17 @@ using Content.Shared.Examine;
 using Content.Shared.FixedPoint;
 using Content.Shared.Fluids.Components;
 using Content.Shared.Friction;
-using Content.Shared.Gravity;
-using Content.Shared.Inventory;
 using Content.Shared.Movement.Components;
 using Content.Shared.Movement.Events;
 using Content.Shared.Movement.Systems;
 using Content.Shared.Nutrition.EntitySystems;
 using Content.Shared.Popups;
 using Content.Shared.Slippery;
-using Content.Shared.Standing;
 using Content.Shared.StepTrigger.Components;
 using Content.Shared.StepTrigger.Systems;
 using Robust.Shared.Audio.Systems;
 using Robust.Shared.Containers;
 using Robust.Shared.Map;
-using Robust.Shared.Physics.Components;
-using Robust.Shared.Physics.Events;
 using Robust.Shared.Prototypes;
 using Robust.Shared.Timing;
 
@@ -47,9 +41,6 @@ public abstract partial class SharedPuddleSystem : EntitySystem
     [Dependency] private SpeedModifierContactsSystem _speedModContacts = default!;
     [Dependency] private StepTriggerSystem _stepTrigger = default!;
     [Dependency] private TileFrictionController _tile = default!;
-    [Dependency] private InventorySystem _inventory = default!; // Funky - Clothing stains
-    [Dependency] private StandingStateSystem _standing = default!; // Moff - Clothing stains
-    [Dependency] private SharedGravitySystem _gravity = default!; // Moff - Clothing Stains
 
     [Dependency] private EntityQuery<StepTriggerComponent> _stepTriggerQuery = default!;
     [Dependency] private EntityQuery<ReactiveComponent> _reactiveQuery = default!;
@@ -106,40 +97,6 @@ public abstract partial class SharedPuddleSystem : EntitySystem
 
         TickEvaporation();
     }
-
-    // Moff start - we basically rewrote this function compared to what funky has
-    // Using startcollide rather than onstep, since the onstep is messed with by slippable... its bleak
-    [SubscribeLocalEvent]
-    private void OnStepInPuddle(Entity<PuddleComponent> ent, ref StartCollideEvent args)
-    {
-        // The thing stepping in the puddle. Because I keep forgetting which is which
-        var stepper = args.OtherEntity;
-
-        if (!_solutionContainerSystem.ResolveSolution(ent.Owner, ent.Comp.SolutionName, ref ent.Comp.Solution, out var solution))
-            return;
-
-        if (solution.Volume <= FixedPoint2.Zero)
-            return;
-
-        // Check if its in air... because... if you're not on the ground you don't get spilled on
-        if (TryComp<PhysicsComponent>(stepper, out var physicsComp)
-            && (physicsComp.BodyStatus == BodyStatus.InAir || _gravity.IsWeightless(stepper)))
-            return;
-
-        // Choose le target...
-        // if standing and have shoes, just get it on their shoes
-        EntityUid target;
-        if (_standing.IsDown(stepper)) // on the ground, spill it on them in general
-            target = stepper;
-        else if (_inventory.TryGetSlotEntity(stepper, "shoes", out var shoes) && shoes is { } shoeUid)
-            target = shoeUid;
-        else
-            return;
-
-        var spilledEvent = new SpilledOnEvent(ent.Owner, solution);
-        RaiseLocalEvent(target, spilledEvent);
-    }
-    // Moff end
 
     private void OnPrototypesReloaded(PrototypesReloadedEventArgs ev)
     {
